+++ gen-enums.config.ts
@@ -1,0 +1,8 @@
+export default {
+  connectionString: "postgres://user:pass@localhost:5432/testdb",
+  schema: "public",
+  outServer: "test/.results-enums/server",
+  outClient: "test/.results-enums/client",
+  softDeleteColumn: null,
+  includeMethodsDepth: 2
+};